requirements.txt
streamlit
langchain
PyPDF2
elevenlabs
pillow
